withdraw logic

+++ crud-bank-app/src/main/java/com/ironhack/crudbankapp/dtos/AmountDTO.java
@@ -3,19 +3,21 @@ package com.ironhack.crudbankapp.dtos;
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
 
+import java.math.BigDecimal;
+
 public class AmountDTO {
     @Max(value = 1000000, message = "The amount cannot be more than 1000000 dollars")
     @Min(value = 0, message = "The amount cannot be less than 0 dollars")
-    private Double amount;
+    private BigDecimal amount;
 
     public AmountDTO() {
     }
 
-    public AmountDTO(Double amount) {
+    public AmountDTO(BigDecimal amount) {
         this.amount = amount;
     }
 
-    public Double getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 }
