+++ crud-bank-app/src/main/java/com/ironhack/crudbankapp/service/interfaces/ICheckingAccountService.java
@@ -2,6 +2,8 @@ package com.ironhack.crudbankapp.service.interfaces;
 
 import com.ironhack.crudbankapp.model.CheckingAccount;
 
+import java.math.BigDecimal;
+
 public interface ICheckingAccountService {
     CheckingAccount getCheckingAccountByAccountNumber(Integer accountNumber);
 
@@ -9,5 +11,5 @@ public interface ICheckingAccountService {
 
     void deleteCheckingAccount(Integer accountNumber);
 
-    void transfer(Integer fromId, Integer destinationId, Double amount);
+    void transfer(Integer fromId, Integer destinationId, BigDecimal amount);
 }
